gui: cleanup rename Java package `dev.flang.feeze` as `dev.feeze`
Since we own the domain `feeze.dev` now we can safely do this.
