+++ src/java/dev/feeze/FeezeToolTip.java
@@ -20,12 +20,12 @@ along with this program.  If not, see <http://www.gnu.org/licenses/>
  *
  * Copyright (c) 2025, Tokiwa Software GmbH, Germany
  *
- * Java source code of class dev.flang.feeze.FeezeToolTip
+ * Java source code of class dev.feeze.FeezeToolTip
  *
  *---------------------------------------------------------------------*/
 
 
-package dev.flang.feeze;
+package dev.feeze;
 
 import java.awt.Dimension;
 import java.awt.Font;
