+++ src/java/dev/feeze/SchedulingPanorama.java
@@ -20,12 +20,12 @@ along with this program.  If not, see <http://www.gnu.org/licenses/>
  *
  * Copyright (c) 2025, Tokiwa Software GmbH, Germany
  *
- * Java source code of class dev.flang.feeze.SchedulingPanorama
+ * Java source code of class dev.feeze.SchedulingPanorama
  *
  *---------------------------------------------------------------------*/
 
 
-package dev.flang.feeze;
+package dev.feeze;
 
 import java.awt.Color;
 import java.awt.Dimension;
