+++ src/java/dev/feeze/SystemThread.java
@@ -20,12 +20,12 @@ along with this program.  If not, see <http://www.gnu.org/licenses/>
  *
  * Copyright (c) 2026, Tokiwa Software GmbH, Germany
  *
- * Java source code of class dev.flang.feeze.SystemThread
+ * Java source code of class dev.feeze.SystemThread
  *
  *---------------------------------------------------------------------*/
 
 
-package dev.flang.feeze;
+package dev.feeze;
 
 import java.util.ArrayList;
 import java.util.Arrays;
