+++ src/java/dev/feeze/SystemUser.java
@@ -20,12 +20,12 @@ along with this program.  If not, see <http://www.gnu.org/licenses/>
  *
  * Copyright (c) 2026, Tokiwa Software GmbH, Germany
  *
- * Java source code of class dev.flang.feeze.SystemProcess
+ * Java source code of class dev.feeze.SystemProcess
  *
  *---------------------------------------------------------------------*/
 
 
-package dev.flang.feeze;
+package dev.feeze;
 
 import java.util.ArrayList;
 
@@ -33,39 +33,54 @@ import java.util.ArrayList;
 
 
 /**
- * SystemProcess  represents a process in recorded data
+ * SystemUser  represents a user in recorded data
  *
  * @author Fridtjof Siebert (siebert@tokiwa.software)
  */
-class SystemProcess
+class SystemUser
 {
-  int _pid;
+  final Data _data;
   int _uid;
   String _name;
   int _num;
-  SystemUser _user;
-  ArrayList<SystemThread> _threads = new ArrayList<>();
+  ArrayList<SystemProcess> _processes = new ArrayList<>();
+  volatile CumulativeThread _cumulative = null;
 
-  SystemProcess(int pid,
-                int uid,
-                String name,
-                int num,
-                SystemUser u)
+  SystemUser(Data data,
+             int uid,
+             String name,
+             int num)
   {
-    _pid = pid;
+    _data = data;
     _uid = uid;
     _name = name;
     _num = num;
-    _user = u;
   }
 
-  void addThread(SystemThread t)
+  void addProcess(SystemProcess p)
   {
-    _threads.add(t);
+    _processes.add(p);
+  }
+
+
+  CumulativeThread cumulative()
+  {
+    if (_cumulative == null)
+      {
+        synchronized (this)
+          {
+            if (_cumulative == null)
+              {
+                var c = new CumulativeThread(this);
+                _cumulative = c;
+              }
+          }
+      }
+    return _cumulative;
   }
 
   public String toString()
   {
-    return _name; // +  " (#" + _pid + ")";
+    return ""+_uid+" "+_name;
   }
 }
