+++ src/java/module-info.java
@@ -1,7 +1,7 @@
 module feeze {
   requires java.desktop;
 
-  exports dev.flang.feeze;
+  exports dev.feeze;
   exports dev.flang.swing;
   exports dev.flang.util;
 }
